+++ NEWS.md
@@ -1,5 +1,6 @@
-# mirai 0.11.1.9003 (development)
+# mirai 0.11.1.9004 (development)
 
+* Using dispatcher with tokens, it is now a requirement to call `saisei()` to regenerate URLs after daemons have disconnected after task or time outs.
 * `make_cluster()` specifying only 'url' now succeeds with implied 'n' of one.
 * Fixes `mirai()` specifying a language object by name for '.expr' in R versions 4.0 and earlier. 
 * Fixes regression in 0.11.1 which prevented the correct random seed being set when using dispatcher.